requirements.txt
python-nmap~=0.7.1
requests~=2.32.5